+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@
 * Clientele will now automatically run [black](https://black.readthedocs.io/en/stable/) code formatter once a client is generated or regenerated.
 * Clientele will now generate absolute paths to refer to adjacent files in the generated client, instead of relative paths. This assumes you are running the `clientele` command in the root directory of your project.
 * A lot of documentation and docs strings updates so that code in the generated client is easier to understand.
+* Improved the utility for snake-casing enum keys. Tests added for the functions.
 
 ## 0.6.3